+++ test/fixtures/alarisa/src/Back/Ingress/Human.mjs
@@ -1,6 +1,0 @@
-export default class HumanIngressFixture {
-  constructor() {
-    this.calls = [];
-    this.accept = async (input) => this.calls.push(input);
-  }
-}